requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
mysql-connector-python==8.0.13
protobuf==3.12.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
